LayoutTests/fast/backgrounds/background-position-parsing-2-expected.txt

Test to make sure background-position is parsed correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


background-position with one value
PASS style.backgroundPosition is '70% 50%'
PASS computedStyle.backgroundPosition is '70% 50%'
PASS style.backgroundPosition is '84px 50%'
PASS computedStyle.backgroundPosition is '84px 50%'
PASS style.backgroundPosition is '0% 50%'
PASS computedStyle.backgroundPosition is '0% 50%'
PASS style.backgroundPosition is '100% 50%'
PASS computedStyle.backgroundPosition is '100% 50%'
PASS style.backgroundPosition is '50% 100%'
PASS computedStyle.backgroundPosition is '50% 100%'
PASS style.backgroundPosition is '50% 0%'
PASS computedStyle.backgroundPosition is '50% 0%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
background-position with two values
PASS style.backgroundPosition is '0% 100%'
PASS computedStyle.backgroundPosition is '0% 100%'
PASS style.backgroundPosition is '0% 100%'
PASS computedStyle.backgroundPosition is '0% 100%'
PASS style.backgroundPosition is '100% 100%'
PASS computedStyle.backgroundPosition is '100% 100%'
PASS style.backgroundPosition is '100% 0%'
PASS computedStyle.backgroundPosition is '100% 0%'
PASS style.backgroundPosition is '54px 100%'
PASS computedStyle.backgroundPosition is '54px 100%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '5% 100%'
PASS computedStyle.backgroundPosition is '5% 100%'
PASS style.backgroundPosition is '30pt -20px'
PASS computedStyle.backgroundPosition is '40px -20px'
PASS style.backgroundPosition is '100% 50%'
PASS computedStyle.backgroundPosition is '100% 50%'
PASS style.backgroundPosition is '100% 0px'
PASS computedStyle.backgroundPosition is '100% 0px'
PASS style.backgroundPosition is '100% 50%'
PASS computedStyle.backgroundPosition is '100% 50%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '0% 50%'
PASS computedStyle.backgroundPosition is '0% 50%'
PASS style.backgroundPosition is '-20% 50%'
PASS computedStyle.backgroundPosition is '-20% 50%'
PASS style.backgroundPosition is '100% 0%'
PASS computedStyle.backgroundPosition is '100% 0%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
background-position invalid with one or two values, no change expected
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
background-position with CSS3 comma separator, one or two values
PASS style.backgroundPosition is '50% 50%, 100% 50%'
PASS computedStyle.backgroundPosition is '50% 50%, 100% 50%'
PASS style.backgroundPosition is '50% 0%, 50% 100%'
PASS computedStyle.backgroundPosition is '50% 0%, 50% 100%'
PASS style.backgroundPosition is '100% 50%, 5px 100%'
PASS computedStyle.backgroundPosition is '100% 50%, 5px 100%'
PASS style.backgroundPosition is '0% 0%, 100% 100%'
PASS computedStyle.backgroundPosition is '0% 0%, 100% 100%'
background-position with CSS3 comma separator, with invalid one or two values, no change expected
PASS style.backgroundPosition is '0% 0%, 100% 100%'
PASS computedStyle.backgroundPosition is '0% 0%, 100% 100%'
PASS style.backgroundPosition is '0% 0%, 100% 100%'
PASS computedStyle.backgroundPosition is '0% 0%, 100% 100%'
PASS style.backgroundPosition is '0% 0%, 100% 100%'
PASS computedStyle.backgroundPosition is '0% 0%, 100% 100%'
background-position with CSS3 four values
PASS style.backgroundPosition is 'left 10px top 15px'
PASS style.backgroundPosition is 'left 10% top 30%'
PASS style.backgroundPosition is 'left 30% top 10%'
PASS style.backgroundPosition is 'right 0% top 15px'
PASS style.backgroundPosition is 'left 10px top 50%'
PASS style.backgroundPosition is 'left 50% top 20px'
PASS style.backgroundPosition is 'left 50% top 20px'
PASS style.backgroundPosition is 'left 30px top 50%'
PASS style.backgroundPosition is 'left 20% top 0%'
background-position with CSS3 invalid four values, no change expected
PASS style.backgroundPosition is 'left 20% top 0%'
PASS style.backgroundPosition is 'left 20% top 0%'
PASS style.backgroundPosition is 'left 20% top 0%'
PASS style.backgroundPosition is 'left 20% top 0%'
PASS style.backgroundPosition is 'left 20% top 0%'
PASS style.backgroundPosition is 'left 20% top 0%'
PASS style.backgroundPosition is 'left 20% top 0%'
PASS style.backgroundPosition is 'left 20% top 0%'
PASS style.backgroundPosition is 'left 20% top 0%'
PASS style.backgroundPosition is 'left 20% top 0%'
PASS style.backgroundPosition is 'left 20% top 0%'
PASS style.backgroundPosition is 'left 20% top 0%'
PASS style.backgroundPosition is 'left 20% top 0%'
PASS style.backgroundPosition is 'left 20% top 0%'
PASS style.backgroundPosition is 'left 20% top 0%'
PASS style.backgroundPosition is 'left 20% top 0%'
PASS style.backgroundPosition is 'left 20% top 0%'
PASS style.backgroundPosition is 'left 20% top 0%'
PASS style.backgroundPosition is 'left 20% top 0%'
PASS style.backgroundPosition is 'left 20% top 0%'
PASS style.backgroundPosition is 'left 20% top 0%'
PASS style.backgroundPosition is 'left 20% top 0%'
PASS style.backgroundPosition is 'left 20% top 0%'
PASS style.backgroundPosition is 'left 20% top 0%'
PASS style.backgroundPosition is 'left 20% top 0%'
PASS style.backgroundPosition is 'left 20% top 0%'
PASS style.backgroundPosition is 'left 20% top 0%'
PASS style.backgroundPosition is 'left 20% top 0%'
PASS style.backgroundPosition is 'left 20% top 0%'
background-position with CSS3 four values and comma
PASS style.backgroundPosition is '50% 50%, left 0% bottom 20px'
PASS style.backgroundPosition is 'right 20px bottom 20px, 0% 50%'
PASS style.backgroundPosition is 'left 10px top 15px, right 20% bottom 20px'
PASS style.backgroundPosition is 'left 10% top 0%, right 0% top 0px'
PASS style.backgroundPosition is 'right 0% top 15px, right 20px bottom 0%'
background-position with invalid CSS3 four values and comma, no change expected
PASS style.backgroundPosition is 'right 0% top 15px, right 20px bottom 0%'
PASS style.backgroundPosition is 'right 0% top 15px, right 20px bottom 0%'
PASS style.backgroundPosition is 'right 0% top 15px, right 20px bottom 0%'
PASS successfullyParsed is true

TEST COMPLETE
